🆙 update : modify menu component
